[RD-3495]: Added separate getter for the installation secrets
